Revert "Merge remote-tracking branch 'origin/sycl' into IRScatter"
This reverts commit 575fd3737b15201bdb5c94518cff331cb249ad2c.

+++ clang/lib/AST/AttrDocTable.cpp
@@ -21,7 +21,7 @@ static const llvm::StringRef AttrDoc[] = {
 };
 
 llvm::StringRef clang::Attr::getDocumentation(clang::attr::Kind K) {
-  if (K < (int)std::size(AttrDoc))
+  if (K < std::size(AttrDoc))
     return AttrDoc[K];
   return "";
 }
